update to 20170429_1(fix bugs)
紧急修复：选择修改标签时导致标签直接丢失
紧急修复：可能的远程执行漏洞

+++ php/init.php
@@ -1,4 +1,5 @@
 <?
+if(!defined('MANAGEversion')){header("HTTP/1.0 404 Not Found");exit("<!DOCTYPE HTML PUBLIC \"-//IETF//DTD HTML 2.0//EN\">\n<html><head>\n<title>404 Not Found</title>\n</head><body>\n<h1>Not Found</h1>\n<p>The requested URL /php/init.php was not found on this server.</p>\n</body></html>\n");}
 
 if($kv->get('tinywebdbMANAGE_password')!=''){exit('<script>window.location.href="?a=index";'.((isset($_GET['de']))?($kv->delete(''.$_GET['de'])):'').'</script>');}
 if(isset($_REQUEST['pwd'])){
